Restructuring prior to Vehicle in progress
