+++ BlueDeck/BlueDeck.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>C:\Users\jcsmith1\source\repos\BlueDeck\BlueDeck\BlueDeck.xml</DocumentationFile>
+    <DocumentationFile>C:\Users\jcs30\source\repos\BlueDeck\BlueDeck\BlueDeck.xml</DocumentationFile>
   </PropertyGroup>
 
   <ItemGroup>
