+++ BlueDeck/Controllers/EnumsApiController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using BlueDeck.Models.APIModels;
 using BlueDeck.Models;
 using Microsoft.AspNetCore.Authorization;
+using BlueDeck.Models.Repositories;
 
 namespace BlueDeck.Controllers
 {
