+++ BlueDeck/Controllers/HomeController.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
+using BlueDeck.Models.Repositories;
 
 namespace BlueDeck.Controllers
 {
