+++ BlueDeck/Controllers/RosterController.cs
@@ -8,6 +8,7 @@ using BlueDeck.Models.Types;
 using System;
 using Microsoft.AspNetCore.Authorization;
 using Newtonsoft.Json;
+using BlueDeck.Models.Repositories;
 
 namespace BlueDeck.Controllers
 {
