+++ BlueDeck/Models/APIModels/GenderApiResult.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using BlueDeck.Models.Enums;
 
 namespace BlueDeck.Models.APIModels
 {
