+++ BlueDeck/Models/APIModels/RoleTypeApiResult.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using BlueDeck.Models.Enums;
 
 namespace BlueDeck.Models.APIModels
 {
