+++ BlueDeck/Models/ApplicationDbContext.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using BlueDeck.Models.Types;
-using System.Collections.Generic;
+using BlueDeck.Models.Enums;
+
 
 namespace BlueDeck.Models {
