+++ BlueDeck/Models/ClaimsLoader.cs
@@ -1,10 +1,13 @@
-﻿using System.Collections.Generic;
+﻿
+using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authentication;
 using Newtonsoft.Json;
 using BlueDeck.Models.Types;
+using BlueDeck.Models.Repositories;
+
 
 namespace BlueDeck.Models
 {
