+++ BlueDeck/Models/Repositories/IPhoneNumberTypeRepository.cs
@@ -1,5 +1,6 @@
 ﻿using BlueDeck.Models.Types;
 using System.Collections.Generic;
+using BlueDeck.Models.Enums;
 
 namespace BlueDeck.Models.Repositories
 {
