+++ BlueDeck/Models/Types/ChartableComponent.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace BlueDeck.Models
+namespace BlueDeck.Models.Types
 {
 
     /// <summary>
